+++ whatsonpypi/cli.py
@@ -58,7 +58,16 @@ from .whatsonpypi import get_query_response
     show_default=True,
     help="Directory to search for requirement files. Only used when --add is used."
 )
-def main(package, more, docs, add, req_dir):
+@click.option(
+    '-p',
+    '--req-pattern',
+    type=str,
+    required=True,
+    default="requirements*.txt",
+    show_default=True,
+    help="Filename pattern for searching requirements files."
+)
+def main(package, more, docs, add, req_dir, req_pattern):
     """
     CLI tool to find the latest version of a package on PyPI.
 
@@ -76,9 +85,10 @@ def main(package, more, docs, add, req_dir):
             more_out=more,
             launch_docs=docs,
             add_to_req=add,
-            req_dir=req_dir
+            req_dir=req_dir,
+            req_pattern=req_pattern
         )
-        # output is not always expected and maybe None sometimes.
+        # output is not always expected and might be None sometimes.
         if output:
             pretty(output)
     except Exception as e: